refactor: reorganize ad state management into separate modules and enhance state handling
