+++ src/content/ads/state.ts
@@ -5,14 +5,6 @@ export type AdState =
 
 export type AdIntent = 'stay' | 'startMute' | 'stopAndUnmute' | 'reset'
 
-/**
- * Decide what the next intent should be based on the current state,
- * whether an ad is detected, and whether the feature is enabled.
- * @param state The current ad state.
- * @param adDetected Whether an ad is currently detected.
- * @param enabled Whether the mute ads feature is enabled.
- * @returns The next ad intent.
- */
 export function decideIntent(
   state: AdState,
   adDetected: boolean,
