+++ src/content/ads/stats.ts
@@ -1,6 +1,6 @@
-import type { AdMuteStats } from '../types'
-import { AD_MUTE_STATS_KEY } from '../types'
-import { logger } from '../utils/logger'
+import type { AdMuteStats } from '../../types'
+import { AD_MUTE_STATS_KEY } from '../../types'
+import { logger } from '../../utils/logger'
 
 export async function recordMutedAd(
   channel: string | null,
